refactor(dicom-web): move components, stores, utils into typed folders
